Clear integration suite PHP warnings and deprecations (#50)
The integration tests render the material-blue theme views, which accessed
nullable/undefined fields and passed them to htmlspecialchars() — producing
'Undefined property/array key' warnings plus 'htmlspecialchars(): Passing
null' deprecations on the same expressions.

- Theme views (.inc): guard the flagged accesses with ?? '' / ?? [] (and an
  opcache_get_status/SERVER_SOFTWARE guard in info.inc). Output is unchanged
  (htmlspecialchars(null) already rendered empty), so the BodyChecker
  assertions still hold.
- A few point-of-use guards for PHP 8.4 deprecations: Account (undefined
  'num' key), AccountSearchHelper ((int)microtime), FileHandler (explicit
  fgetcsv enclosure/escape), Serde (@unserialize for fixtures with trailing
  bytes; the prefix value is still returned).

Integration Warnings 39 -> 0, Deprecations 30 -> 0 (the 1 FileHandler notice
is separate). Integration 93/93, unit 1981/1981.

+++ app/modules/web/Controllers/Helpers/Account/AccountSearchHelper.php
@@ -86,7 +86,7 @@ final class AccountSearchHelper extends HelperBase
     ) {
         parent::__construct($application, $template, $request);
 
-        $this->queryTimeStart = microtime(true);
+        $this->queryTimeStart = (int)microtime(true);
         $this->isIndex = $this->request->analyzeString('r') ===
                          $this->acl->getRouteFor(AclActionsInterface::ACCOUNT);
         $this->setVars();
